Fix issue in case of empty report fields
The current version of the mainline zap-friendly-report encounters
problems if it tries to load a ZAP report file that has some empty
fields. As this was the case for some scans that I was performing, I
created a simple patch to make it work.

This patch has some cosmetic side effects, which I did not have the
patience to track down and fix. Use at your own risk :)

Signed-off-by: Max Maass <max.maass@iteratec.com>